+++ vite.config.js
@@ -1,11 +1,20 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
+import Markdown from "unplugin-vue-markdown/vite";
 
-import markdownInVue from "./index.js";
+
+import MarkdownInVue from "./index.js";
 
 export default defineConfig({
   plugins: [
-    markdownInVue(),
-    vue(),
+    MarkdownInVue(),
+    MarkdownInVue({
+      elementNameBlock: "MdBlock",
+      elementNameInline: "MdInline"
+    }),
+    Markdown(),
+    vue({
+      include: [/\.vue$/, /\.md$/]
+    }),
   ],
 })